changed file names
modified web_controller to be new main.py
